prevent some specific tests from running in parallel
