+++ src/Microsoft.DotNet.Interactive.Tests/LanguageKernelTestBase.cs
@@ -15,7 +15,6 @@ using Microsoft.DotNet.Interactive.PowerShell;
 using Microsoft.DotNet.Interactive.Tests.Utility;
 
 using Pocket;
-using Pocket.For.Xunit;
 
 using static Pocket.Logger<Microsoft.DotNet.Interactive.Tests.LanguageKernelTestBase>;
